Removed old shadow code and small refactor
