+++ app/src/main/java/com/goldenraven/padawanwallet/theme/Modifiers.kt
@@ -1,6 +1,16 @@
 package com.goldenraven.padawanwallet.theme
 
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
 import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Paint
@@ -8,6 +18,7 @@ import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import com.goldenraven.padawanwallet.utils.BackgroundShape
 
 fun Modifier.advancedShadow(
     color: Color = Color.Black,
@@ -40,5 +51,39 @@ fun Modifier.advancedShadow(
             cornersRadius.toPx(),
             paint
         )
+    }
+}
+
+fun Modifier.standardBackground(): Modifier = this.then(
+    fillMaxSize()
+        .background(padawan_theme_background_secondary, shape = BackgroundShape())
+        .padding(all = 32.dp)
+)
+
+fun Modifier.wideTextField(): Modifier = this.then(
+    Modifier
+        .padding(vertical = 8.dp)
+        .fillMaxWidth()
+        .border(
+            width = 2.dp,
+            color = padawan_theme_onPrimary,
+            shape = RoundedCornerShape(20.dp)
+        )
+        .advancedShadow(
+            color = Color.Black,
+            alpha = 1f,
+            cornersRadius = 20.dp,
+            shadowBlurRadius = 0.0001.dp,
+            offsetX = 4.dp,
+            offsetY = 4.dp
+        )
+)
+
+inline fun Modifier.noRippleClickable(crossinline onClick: () -> Unit): Modifier = composed {
+    clickable(
+        indication = null,
+        interactionSource = remember { MutableInteractionSource() }
+    ) {
+        onClick()
     }
 }
